+++ changelog.md
@@ -1,4 +1,8 @@
 [![Downloads](https://img.shields.io/github/downloads/probablyraging/steam-game-idler/1.5.21/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/probablyraging/steam-game-idler/releases/download/1.5.21/Steam.Game.Idler_1.5.21_x64_en-US.msi)
 
 ## Changelog
-- Fixed an issue caused by the recent Steam client update, which prevented achievements from being unlocked
+- Show a error screen when the games list is unable to be retrieved
+  - This usually occurs when:
+    - Your profile is set to private
+    - Your game details are set to private
+    - You have no games in your library